+++ MovieApp/MovieApp/App/UseCase/Model/AuthorModel.swift
@@ -5,7 +5,7 @@ struct AuthorModel {
     let name: String
     let username: String
     let avatarPath: URL?
-    let rating: Int?
+    let rating: Double?
     
 }
